Moved response tests to separate files, added provided tests for response.
