+++ src/test/helpers/response.ts
@@ -1,0 +1,47 @@
+// Types
+import type { Socket } from 'net'
+import type { Response } from 'express'
+
+export function callAllFunctions(res: Response) {
+  // express - Response
+  res.status(123)
+  res.sendStatus(123)
+  res.links({})
+  res.send()
+  res.json()
+  res.jsonp()
+  res.sendFile('test')
+  res.sendfile('test')
+  res.download('test')
+  res.contentType('test')
+  res.type('test')
+  res.format({})
+  res.attachment()
+  res.set({})
+  res.header({})
+  res.get('test')
+  res.clearCookie('test')
+  res.cookie('test', 'test')
+  res.location('test')
+  res.redirect('test')
+  res.render('test')
+  res.vary('test')
+  res.append('test')
+  // http - ServerResponse
+  res.end()
+  res.assignSocket({} as Socket)
+  res.detachSocket({} as Socket)
+  res.writeContinue()
+  res.writeHead(123)
+  res.writeProcessing()
+  // http - OutgoingMessage
+  res.setTimeout(123)
+  res.setHeader('test', 'test')
+  res.getHeader('test')
+  res.getHeaders()
+  res.getHeaderNames()
+  res.hasHeader('test')
+  res.removeHeader('test')
+  res.addTrailers([])
+  res.flushHeaders()
+}
